add santa hat
